develop into deploy

See merge request s12-bigdata-recom-sub1/S12P21C104!135

+++ backend/src/main/java/com/ssafy/backend/ai/sound/dto/request/SoundGenerationRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class SoundGenerationRequest {
     private String genre; // 음악 장르
     private String mood; // 분위기
-    private int durationSeconds; // 음원 길이(초)
+    private double durationSeconds; // 음원 길이(초)
     private int tempo; // 템포(BPM)
     private String instruments; // 사용할 악기(콤마로 구분)
     private String additionalNotes; // 추가 요청사항
